requirements.txt
python-dotenv
langchain-openai
streamlit
langchain_community